sorely needed refactor into plugins
